pyproject.toml
[tool.poetry]
name = "ab-admin"
version = "0.1.0"
description = "Discord bot for to manage archive bot community discord server."
authors = ["Rory <rory@cryptii.co.uk>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
"discord.py" = "^2.4.0"
certifi = "^2024.8.30"
python-dotenv = "^1.0.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
